Fix corner cases in -mibt-preceding-endbr
While compiling the kernel/running it with QEMU, we noticed that some
indirect call conversions made by -mibt-preceding-endbr were clobbering
some function arguments, despite the liveness analysis used to verify if
it was safe to use the registers. This was mostly because we were only
looking for the "isKill" flag, which means that the register is dead in
the next instruction... but sometimes these registers were actually very
alive in the functions that were being called (because they are used to
pass parameters).

Since there is always a MOV instruction when CALL64m is being converted,
we already emit this to a free register and then use this register to do
the call, this way we'll never need to restore it after the call.

For the CALL64r the MOV is only needed when the register used for the call
is also used as an argument. I haven't seen this case in practice, but I
think it could happen... the fix implemented should handle it, but is not
properly tested.

Note: while hunting the corner cases, I tried conservatively always
restoring the register value of a CALL64r right after it, resulting in the
following sequence:

sub 4, reg
call *reg
add 4, reg

But here an interesting situation appeared -- when using RAX to call, this
would destroy the return value. Thus, being this conservative is not an
option and the "isKill()" analysis is mandatory.